変数をfloatに変更したことで不要になった処理を削除

+++ Assets/_WorkSpace/Script/Player/HPBar.cs
@@ -16,7 +16,7 @@ public class HPBar : MonoBehaviour
 
     void Update()
     {
-        _hpBarImage.fillAmount = (float)_pH._playerHealth / 100f;
+        _hpBarImage.fillAmount = _pH._playerHealth / 100f;
         Debug.Log($"{_pH._playerHealth / 100}");
     }
 }
